+++ modules/portal-jslib/src/test/java/com/enonic/xp/portal/jslib/impl/ImageUrlHandlerTest.java
@@ -2,14 +2,11 @@ package com.enonic.xp.portal.jslib.impl;
 
 import org.junit.Test;
 
-import com.enonic.wem.script.command.CommandHandler;
-import com.enonic.xp.portal.jslib.AbstractHandlerTest;
-
 public class ImageUrlHandlerTest
-    extends AbstractHandlerTest
+    extends AbstractUrlHandlerTest
 {
     @Override
-    protected CommandHandler createHandler()
+    protected AbstractUrlHandler createUrlHandler()
     {
         return new ImageUrlHandler();
     }